code_demo/requirements.txt
alabaster==1.0.0
annotated-types==0.7.0
app-model==0.2.8
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
brokenaxes==0.6.2
build==1.2.2
cachey==0.2.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
dask==2024.9.0
debugpy==1.6.7
decorator==5.1.1
docstring_parser==0.16
docutils==0.21.2
einops==0.6.1
executing==1.2.0
flexcache==0.3
flexparser==0.3.1
fonttools==4.39.3
freetype-py==2.5.1
fsspec==2024.9.0
HeapDict==1.0.1
hsluv==5.0.4
idna==3.10
imageio==2.28.1
imagesize==1.4.1
importlib_metadata==8.5.0
in-n-out==0.2.1
ipykernel==6.22.0
ipython==8.13.1
ipywidgets==8.1.5
jedi==0.18.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.4
lazy_loader==0.2
locket==1.0.0
magicgui==0.9.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
napari==0.5.3
napari-console==0.0.9
napari-plugin-engine==0.2.0
napari-svg==0.2.0
nest-asyncio==1.5.6
networkx==3.1
npe2==0.7.7
numpy==1.24.3
numpydoc==1.8.0
packaging==23.1
pandas==2.2.3
parso==0.8.3
partd==1.4.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
Pint==0.24.3
platformdirs==3.5.0
pooch==1.8.2
prompt-toolkit==3.0.38
psutil==5.9.5
psygnal==0.11.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyconify==0.1.6
pydantic==2.9.2
pydantic-compat==0.1.2
pydantic_core==2.23.4
Pygments==2.18.0
PyOpenGL==3.1.7
pyparsing==3.0.9
pyproject_hooks==1.1.0
python-dateutil==2.8.2
pytz==2024.2
pyvista==0.44.1
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==25.0.2
qtconsole==5.6.0
QtPy==2.4.1
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
scikit-image==0.20.0
scipy==1.10.1
scooby==0.10.0
shellingham==1.5.4
six==1.16.0
snowballstemmer==2.2.0
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.2
superqt==0.6.7
tabulate==0.9.0
tifffile==2023.4.12
tomli_w==1.0.0
toolz==0.12.1
tornado==6.3.1
tqdm==4.66.5
traitlets==5.9.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vispy==0.14.3
vtk==9.3.1
wcwidth==0.2.6
widgetsnbextension==4.0.13
wrapt==1.16.0
zipp==3.20.2
